stylize the tsci dev command (#112)
* stylize the tsci dev command

* make kleur a dev dep

* remove check on dependencies

* separate the pcb svg export

* split export test into separate files

* format test files
