+++ lib/server/createHttpServer.ts
@@ -11,6 +11,7 @@ import runFrameStandaloneBundleContent from "@tscircuit/runframe/standalone" wit
 // @ts-ignore
 import winterspecBundle from "@tscircuit/file-server/dist/bundle.js"
 import { getIndex } from "../site/getIndex"
+import kleur from "kleur"
 
 export const createHttpServer = async (port = 3020) => {
   const fileServerHandler = getNodeHandler(winterspecBundle as any, {})
@@ -43,7 +44,7 @@ export const createHttpServer = async (port = 3020) => {
       }
 
       res.writeHead(302, {
-        Location: `https://cdn.jsdelivr.net/npm/@tscircuit/runframe@${{ ...pkg.devDependencies, ...pkg.dependencies }["@tscircuit/runframe"].replace(/^[^0-9]+/, "")}/dist/standalone.min.js`,
+        Location: `https://cdn.jsdelivr.net/npm/@tscircuit/runframe@${{ ...pkg.devDependencies }["@tscircuit/runframe"].replace(/^[^0-9]+/, "")}/dist/standalone.min.js`,
       })
       res.end()
       return
@@ -68,7 +69,6 @@ export const createHttpServer = async (port = 3020) => {
 
   return new Promise<{ server: http.Server }>((resolve) => {
     server.listen(port, () => {
-      console.log(`Server running at http://localhost:${port}`)
       resolve({ server })
     })
   })
